Quick and dirty fix for a caudium backtrace.
Some modules (for example relay.pike) don't provide the last funtion to
shuffle(object from, object to,object|void to2, function(:void)|void callback).
So the callback function is void in shuffle and so there is an error when doing
p->set_done_callback(callback);
In caudium 1.0, this code is between #if constant(spider) and so is never
executed and so it works. I made the same thing here but i think someone should
investigate.


git-svn-id: https://caudium.googlecode.com/svn/caudium/branches/stable_1_2@1991 d22b66b4-c94e-0410-8570-a1be9e963a58