why not have the same "docs" in there.
besides, i should add that it's very dirty, so beware. i encountered a number
of oddities wrt combine_path() and the value of `f'. i'm not pretending that
i fully understand the thing everywhere, it just seems to work. for what
bertrand needs it, probably a whole lot of cruft can be removed from
find_dir()/find_file() as well, but i tried to keep the changes (compared
to filesystem.pike) as minimal as possible.


git-svn-id: https://caudium.googlecode.com/svn/caudium/trunk@2679 d22b66b4-c94e-0410-8570-a1be9e963a58