first cut at PSP (pike server pages)
syntax:

<%= someexpression %>  <-- gives you some expression cast as a string.

<%  somepikecode %>  <-- inserts some pike code fragments in page

the object "out" is available and is a String.Buffer. All of your output should be `+=() to this object.

return doesn't play a role here; don't use it!

blocks of non-psp markup are considered a block, so you can do this:

<% if(id->variables->somevar) %>
all this
appears if somevar is present and non-zero.
<% else %>
all of this appears otherwise, until we hit another <% %> (ie this shows up
regardless) block.

note that you can use {} in these cases to force the situation, as all text blocks get converted automatically to output().


git-svn-id: https://caudium.googlecode.com/svn/caudium/trunk@5404 d22b66b4-c94e-0410-8570-a1be9e963a58